requirements.txt
langchain
python-dotenv
elevenlabs
Flask
playsound
openai
PyObjC